Remove unnecessary static_casts

+++ tests/UnitTests/ImageOperations/TestSources/EchoImage.cpp
@@ -24,7 +24,7 @@ LIBRARY_LINK_FUNCTION(EchoImage1) {
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
 
 LIBRARY_LINK_FUNCTION(EchoImage2) {
@@ -63,7 +63,7 @@ LIBRARY_LINK_FUNCTION(EchoImage2) {
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
 
 LIBRARY_LINK_FUNCTION(ConvertImageToByte) {
@@ -81,7 +81,7 @@ LIBRARY_LINK_FUNCTION(ConvertImageToByte) {
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
 
 LIBRARY_LINK_FUNCTION(UnifyImageTypes) {
@@ -104,5 +104,5 @@ LIBRARY_LINK_FUNCTION(UnifyImageTypes) {
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
